+++ packages/core/stories/flex-layout/flex-layout.stories.tsx
@@ -14,7 +14,7 @@ export default {
       control: { type: "select" },
     },
     direction: {
-      options: ["row", "column"],
+      options: ["row", "column", "row-reverse", "column-reverse"],
       control: { type: "select" },
     },
     gap: {