+++ packages/embla-carousel/src/CarouselSlides.tsx
@@ -1,18 +1,31 @@
-import { makePrefixer, useForkRef } from "@salt-ds/core";
+import { makePrefixer, useAriaAnnouncer, useForkRef } from "@salt-ds/core";
 import { useComponentCssInjection } from "@salt-ds/styles";
 import { useWindow } from "@salt-ds/window";
 import { clsx } from "clsx";
 import type { EmblaCarouselType } from "embla-carousel";
 import {
+  Children,
   type ComponentPropsWithoutRef,
+  cloneElement,
+  type FocusEvent,
   forwardRef,
   type KeyboardEvent,
+  type MouseEventHandler,
+  type ReactElement,
   useEffect,
   useRef,
+  useState,
 } from "react";
-import { useCarouselContext } from "./CarouselContext";
+import {
+  type CarouselAnnouncementTrigger,
+  useCarouselContext,
+} from "./CarouselContext";
 import carouselSlidesCss from "./CarouselSlides.css";
 import { createCustomSettle } from "./createCustomSettle";
+import { getVisibleSlideDescription } from "./getVisibleSlideDescription";
+import { getVisibleSlideIndexes } from "./getVisibleSlideIndexes";
+
+const ANNOUNCEMENT_DURATION = 1200;
 
 /**
  * Props for the CarouselSlides component.
@@ -21,78 +34,218 @@ export interface CarouselSlidesProps extends ComponentPropsWithoutRef<"div"> {}
 
 const withBaseName = makePrefixer("saltCarouselSlides");
 
+const announceSlideChangesFrom: CarouselAnnouncementTrigger[] = [
+  "drag",
+  "focus",
+  "navigation",
+];
+
 export const CarouselSlides = forwardRef<HTMLDivElement, CarouselSlidesProps>(
-  function CarouselSlides({ children, className, onKeyDown, ...rest }, ref) {
+  function CarouselSlides(
+    { children, className, id, onKeyDown, ...rest },
+    ref,
+  ) {
     const targetWindow = useWindow();
     useComponentCssInjection({
       testId: "salt-carousel-slides",
       css: carouselSlidesCss,
       window: targetWindow,
     });
-    const { emblaApi, emblaRef } = useCarouselContext();
+    const {
+      disableSlideAnnouncements,
+      announcementState,
+      setAnnouncementState,
+      emblaApi,
+      emblaRef,
+      carouselId,
+    } = useCarouselContext();
 
-    const carouselRef = useForkRef<HTMLDivElement>(ref, emblaRef);
+    const containerRef = useRef<HTMLDivElement>(null);
+    const forkedEmblaRef = useForkRef<HTMLDivElement>(ref, emblaRef);
+    const carouselRef = useForkRef<HTMLDivElement>(
+      forkedEmblaRef,
+      containerRef,
+    );
 
-    const usingArrowNavigation = useRef<boolean>();
+    const slideRefs = useRef<(HTMLDivElement | null)[]>([]);
+    const [focusedSlideIndex, setFocusedSlideIndex] = useState<number>(-1);
+    const [dragging, setDragging] = useState(false);
+    const focusOnSettle = useRef<boolean>(false);
+
+    const [stableScrollSnap, setStableScrollSnap] = useState<
+      number | undefined
+    >(undefined);
+
+    const visibleSlideIndexes = getVisibleSlideIndexes(
+      emblaApi,
+      stableScrollSnap ?? 0,
+    );
+    const { announce } = useAriaAnnouncer();
 
     useEffect(() => {
       const handleSettle = (emblaApi: EmblaCarouselType) => {
-        if (!usingArrowNavigation.current) {
-          return;
-        }
-        const slideIndexInView = emblaApi?.selectedScrollSnap() ?? 0;
-        const snappedSlide = emblaApi.slideNodes()[slideIndexInView];
-        if (snappedSlide) {
-          const focusableElements = snappedSlide.querySelectorAll<HTMLElement>(
-            'a, button, input, textarea, select, [tabindex]:not([tabindex="-1"])',
-          );
-          if (focusableElements.length > 0) {
-            focusableElements[0].focus();
-          }
-        }
-        usingArrowNavigation.current = false;
+        const selectedScrollSnap = emblaApi?.selectedScrollSnap() ?? 0;
+        setStableScrollSnap(selectedScrollSnap);
+        const numberOfSnaps = emblaApi?.scrollSnapList().length ?? 1;
+        const numberOfSlidesPerSnap = slideRefs.current.length / numberOfSnaps;
+        const settledSlideIndex = Math.floor(
+          selectedScrollSnap * numberOfSlidesPerSnap,
+        );
+        setFocusedSlideIndex(settledSlideIndex);
       };
 
       if (!emblaApi) {
         return;
       }
+
       const scrollCallback = createCustomSettle(handleSettle);
+      const pointerDownCallback = () => {
+        setAnnouncementState("drag");
+      };
       emblaApi.on("scroll", scrollCallback);
+      emblaApi.on("pointerDown", pointerDownCallback);
       // Cleanup listener on component unmount
       return () => {
         emblaApi.off("scroll", scrollCallback);
+        emblaApi.off("pointerDown", pointerDownCallback);
       };
-    }, [emblaApi]);
+    }, [emblaApi, setAnnouncementState]);
+
+    useEffect(() => {
+      if (!focusOnSettle.current || stableScrollSnap === undefined) {
+        return;
+      }
+      slideRefs.current[focusedSlideIndex]?.focus();
+      setAnnouncementState("focus");
+      focusOnSettle.current = false;
+    }, [stableScrollSnap, focusedSlideIndex, setAnnouncementState]);
+
+    useEffect(() => {
+      const numberOfSnaps = emblaApi?.scrollSnapList().length ?? 1;
+      const numberOfSlidesPerSnap = slideRefs.current.length / numberOfSnaps;
+      if (focusedSlideIndex >= 0) {
+        const nearestScrollSnap = Math.floor(
+          focusedSlideIndex / numberOfSlidesPerSnap,
+        );
+        if (emblaApi?.selectedScrollSnap() !== nearestScrollSnap) {
+          emblaApi?.scrollTo(nearestScrollSnap);
+          focusOnSettle.current = true;
+        }
+      } else if (focusedSlideIndex === -1) {
+        const initialSnap = emblaApi?.selectedScrollSnap();
+        const initialSlideIndex =
+          initialSnap !== undefined
+            ? Math.floor(initialSnap * numberOfSlidesPerSnap)
+            : 0;
+        setFocusedSlideIndex(initialSlideIndex);
+        setStableScrollSnap(initialSnap);
+      }
+    }, [focusedSlideIndex, emblaApi]);
+
+    useEffect(() => {
+      if (disableSlideAnnouncements === false) {
+        setAnnouncementState(undefined);
+      }
+    }, [disableSlideAnnouncements, setAnnouncementState]);
+
+    useEffect(() => {
+      if (
+        stableScrollSnap === undefined ||
+        disableSlideAnnouncements ||
+        !announcementState ||
+        announceSlideChangesFrom.indexOf(announcementState) === -1
+      ) {
+        return;
+      }
+      const announcement = getVisibleSlideDescription(
+        emblaApi,
+        stableScrollSnap,
+      );
+      announce(announcement, ANNOUNCEMENT_DURATION);
+    }, [
+      announce,
+      announcementState,
+      disableSlideAnnouncements,
+      stableScrollSnap,
+      emblaApi,
+    ]);
 
     const handleKeyDown = (event: KeyboardEvent<HTMLDivElement>) => {
       if (event.repeat) {
         return;
       }
+      const numberOfSnaps = emblaApi?.scrollSnapList().length ?? 1;
+      const numberOfSlidesPerSnap = slideRefs.current.length / numberOfSnaps;
+
+      // Find the current snap
+      const currentSnap = Math.floor(focusedSlideIndex / numberOfSlidesPerSnap);
+
       switch (event.key) {
         case "ArrowLeft": {
           event.preventDefault();
-          emblaApi?.scrollPrev();
-          usingArrowNavigation.current = true;
+          const prevSnap = Math.max(currentSnap - 1, 0);
+          const newIndex = prevSnap * numberOfSlidesPerSnap;
+          setFocusedSlideIndex(newIndex);
           break;
         }
         case "ArrowRight": {
           event.preventDefault();
-          emblaApi?.scrollNext();
-          usingArrowNavigation.current = true;
+          const nextSnap = Math.min(currentSnap + 1, numberOfSnaps - 1);
+          const newIndex = nextSnap * numberOfSlidesPerSnap;
+          setFocusedSlideIndex(newIndex);
           break;
         }
       }
       onKeyDown?.(event);
     };
 
+    const handleMouseDown: MouseEventHandler<HTMLDivElement> = (event) => {
+      setDragging(true);
+      rest.onMouseDown?.(event);
+    };
+    const handleMouseUp: MouseEventHandler<HTMLDivElement> = (event) => {
+      setDragging(false);
+      rest.onMouseUp?.(event);
+    };
+
     return (
       <div
         onKeyDown={handleKeyDown}
+        onMouseDown={handleMouseDown}
+        onMouseUp={handleMouseUp}
         ref={carouselRef}
-        className={clsx(withBaseName(), className)}
+        className={clsx(
+          withBaseName(),
+          { [withBaseName("dragging")]: dragging },
+          className,
+        )}
         {...rest}
       >
-        <div className={withBaseName("container")}>{children}</div>
+        <div
+          className={withBaseName("container")}
+          id={id ?? `${carouselId}-slides`}
+        >
+          {Children.map(children, (child, index) => {
+            const childElement = child as ReactElement;
+            const existingId = childElement.props.id;
+            const isFocused = focusedSlideIndex === index;
+            const isHidden =
+              !visibleSlideIndexes.includes(index + 1) && !isFocused;
+            const element = child as ReactElement;
+            return cloneElement(element, {
+              "aria-hidden": isHidden,
+              id: existingId ?? `${carouselId}-slide${index + 1}`,
+              onFocus: (event: FocusEvent) => {
+                setFocusedSlideIndex(index);
+                element.props?.onFocus?.(event);
+              },
+              tabIndex: !isHidden ? 0 : -1,
+              ref: (el: HTMLDivElement) => {
+                slideRefs.current[index] = el;
+              },
+            });
+          })}
+        </div>
       </div>
     );
   },