+++ site/src/examples/number-input/Bordered.tsx
@@ -1,5 +1,12 @@
 import { NumberInput } from "@salt-ds/lab";
 
 export const Bordered = () => (
-  <NumberInput defaultValue={0} bordered style={{ width: "256px" }} />
+  <NumberInput
+    defaultValue={0}
+    bordered
+    onNumberChange={(_event, newValue) =>
+      console.log(`Number changed to ${newValue}`)
+    }
+    style={{ width: "256px" }}
+  />
 );